Rewrite distance calculation in C# to improve performance considerably.
Distance calculation was originally written in C++. It run as an independent process and communicated with the C# program using pipes. The communication overhead was huge and it became the performance bottleneck.

This commit rewrites distance calculation in C# to avoid the overhead. The rewritten class is ugly because it's almost directly copy-pasted from the origianal C++ code. This is fine since the the rewritten code works in exactly the same way as the original code and no change is going to be made to it.